requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.0.0

# Optional: Progress bars
tqdm>=4.62.0

# Rendering (optional, only needed for image rendering)
# gsplat  # Install separately if using local GS rendering
